[SP-3314] - Backport of CDA-91 - As a Developer I can't invoke Jobs or Transformations from "Initial Transformation" (7.0 Suite)  - revert kettle part of the fix since causes a regreassion
