+++ engine/src/org/pentaho/di/trans/messages/messages_en_US.properties
@@ -292,4 +292,3 @@ Trans.Log.DispacthingStartedForFilename=Dispatching started for filename [{0}]
 TransMeta.MissingPluginsFoundWhileLoadingTransformation.Exception=Missing plugins found while loading a transformation
 Trans.FinishListeners.Exception=Error running finish transformation listners
 TransSplitter.Clustering.CopyNumberStep=The number of step copies on the master has to be 1 or equal to the number of slaves ({0}) to work. Note that you can insert a dummy step between {1} and {2} steps to make the transformation work as desired.
-StepWithMappingMeta.Exception.UnableToLoadTrans=It was not possible to load the specified transformation
