+++ engine/src/org/pentaho/di/trans/messages/messages_it_IT.properties
@@ -270,4 +270,3 @@ TransMeta.CheckResult.TypeResultError.CannotFindPreviousFields.Description=Impos
 TransMeta.Monitor.SavingNoteTask.Title=Salvataggio nota \#
 TransMeta.Exception.OnlyOneMappingOutputStepAllowed=Senza specificare dei nomi di passo da cui leggere, ci pu\u00F2 essere solo un passo 'mappatura d''output' nella mappatura, ne sono stati trovati {0}.
 Trans.Log.DispacthingStartedForFilename=Spedizione iniziata per il file [{0}]
-StepWithMappingMeta.Exception.UnableToLoadTrans=Impossibile caricare la trasformazione specificata
