+++ engine/src/org/pentaho/di/trans/messages/messages_ja_JP.properties
@@ -284,4 +284,3 @@ PeekMetric.Column.Errors=\u30a8\u30e9\u30fc
 PeekMetric.Column.Rejected=\u62d2\u5426
 PeekMetric.Column.PriorityBufferSizes=\u5165\u529b/\u51fa\u529b
 PeekMetric.Column.Stepname=\u30b9\u30c6\u30c3\u30d7\u540d
-StepWithMappingMeta.Exception.UnableToLoadTrans=It was not possible to load the specified transformation
