+++ engine/src/org/pentaho/di/trans/steps/simplemapping/SimpleMappingMeta.java
@@ -56,10 +56,10 @@ import org.pentaho.di.resource.ResourceEntry;
 import org.pentaho.di.resource.ResourceEntry.ResourceType;
 import org.pentaho.di.resource.ResourceNamingInterface;
 import org.pentaho.di.resource.ResourceReference;
-import org.pentaho.di.trans.StepWithMappingMeta;
 import org.pentaho.di.trans.Trans;
 import org.pentaho.di.trans.TransMeta;
 import org.pentaho.di.trans.TransMeta.TransformationType;
+import org.pentaho.di.trans.step.BaseStepMeta;
 import org.pentaho.di.trans.step.StepDataInterface;
 import org.pentaho.di.trans.step.StepIOMeta;
 import org.pentaho.di.trans.step.StepIOMetaInterface;
@@ -82,9 +82,13 @@ import org.w3c.dom.Node;
  *
  */
 
-public class SimpleMappingMeta extends StepWithMappingMeta implements StepMetaInterface, HasRepositoryInterface {
-
+public class SimpleMappingMeta extends BaseStepMeta implements StepMetaInterface, HasRepositoryInterface {
   private static Class<?> PKG = SimpleMappingMeta.class; // for i18n purposes, needed by Translator2!!
+  private String transName;
+  private String fileName;
+  private String directoryPath;
+  private ObjectId transObjectId;
+  private ObjectLocationSpecificationMethod specificationMethod;
 
   private MappingIODefinition inputMapping;
   private MappingIODefinition outputMapping;
@@ -400,7 +404,94 @@ public class SimpleMappingMeta extends StepWithMappingMeta implements StepMetaIn
     return null;
   }
 
+  public static final synchronized TransMeta loadMappingMeta( SimpleMappingMeta mappingMeta, Repository rep,
+    IMetaStore metaStore, VariableSpace space ) throws KettleException {
+    TransMeta mappingTransMeta = null;
 
+    switch ( mappingMeta.getSpecificationMethod() ) {
+      case FILENAME:
+        String realFilename = space.environmentSubstitute( mappingMeta.getFileName() );
+        try {
+          // OK, load the meta-data from file...
+          //
+          // Don't set internal variables: they belong to the parent thread!
+          //
+          mappingTransMeta = new TransMeta( realFilename, metaStore, rep, true, space, null );
+          mappingTransMeta.getLogChannel().logDetailed(
+            "Loading Mapping from repository",
+            "Mapping transformation was loaded from XML file [" + realFilename + "]" );
+        } catch ( Exception e ) {
+          throw new KettleException( BaseMessages.getString(
+            PKG, "SimpleMappingMeta.Exception.UnableToLoadMapping" ), e );
+        }
+        break;
+
+      case REPOSITORY_BY_NAME:
+        String realTransname = space.environmentSubstitute( mappingMeta.getTransName() );
+        String realDirectory = space.environmentSubstitute( mappingMeta.getDirectoryPath() );
+
+        if ( rep == null ) { // hardening because TransMeta.setRepositoryOnMappingSteps(); might be missing in special
+                             // situations
+          throw new KettleException( BaseMessages.getString(
+            PKG, "SimpleMappingMeta.Exception.InternalErrorRepository.Message" ) );
+        }
+
+        if ( !Utils.isEmpty( realTransname ) && !Utils.isEmpty( realDirectory ) && rep != null ) {
+          RepositoryDirectoryInterface repdir = rep.findDirectory( realDirectory );
+          if ( repdir != null ) {
+            try {
+              // reads the last revision in the repository...
+              //
+              mappingTransMeta = rep.loadTransformation( realTransname, repdir, null, true, null ); // TODO: FIXME:
+                                                                                                    // Should we pass in
+                                                                                                    // external
+                                                                                                    // MetaStore into
+                                                                                                    // Repository
+                                                                                                    // methods?
+
+              mappingTransMeta.getLogChannel().logDetailed(
+                "Loading Mapping from repository",
+                "Mapping transformation [" + realTransname + "] was loaded from the repository" );
+            } catch ( Exception e ) {
+              throw new KettleException( "Unable to load transformation [" + realTransname + "]", e );
+            }
+          } else {
+            throw new KettleException( BaseMessages.getString(
+              PKG, "SimpleMappingMeta.Exception.UnableToLoadTransformation", realTransname )
+              + realDirectory );
+          }
+        } else {
+          throw new KettleException( BaseMessages.getString(
+            PKG, "SimpleMappingMeta.Exception.UnableToLoadTransformationNameOrDirNotGiven" ) );
+        }
+        break;
+
+      case REPOSITORY_BY_REFERENCE:
+        // Read the last revision by reference...
+        if ( rep == null ) { // hardening because TransMeta.setRepositoryOnMappingSteps(); might be missing in special
+                             // situations
+          throw new KettleException( BaseMessages.getString(
+            PKG, "SimpleMappingMeta.Exception.InternalErrorRepository.Message" ) );
+        }
+        mappingTransMeta = rep.loadTransformation( mappingMeta.getTransObjectId(), null );
+        break;
+      default:
+        break;
+    }
+
+    // Pass some important information to the mapping transformation metadata:
+    //
+    if ( mappingTransMeta == null ) { // hardening because TransMeta might have issues in special situations
+      throw new KettleException( BaseMessages.getString(
+        PKG, "SimpleMappingMeta.Exception.InternalErrorTransMetaIsNULL.Message" ) );
+    }
+    mappingTransMeta.copyVariablesFrom( space );
+    mappingTransMeta.setRepository( rep );
+    mappingTransMeta.setMetaStore( metaStore );
+    mappingTransMeta.setFilename( mappingTransMeta.getFilename() );
+
+    return mappingTransMeta;
+  }
 
   public void check( List<CheckResultInterface> remarks, TransMeta transMeta, StepMeta stepMeta,
     RowMetaInterface prev, String[] input, String[] output, RowMetaInterface info, VariableSpace space,
@@ -439,6 +530,51 @@ public class SimpleMappingMeta extends StepWithMappingMeta implements StepMetaIn
 
   public StepDataInterface getStepData() {
     return new SimpleMappingData();
+  }
+
+  /**
+   * @return the directoryPath
+   */
+  public String getDirectoryPath() {
+    return directoryPath;
+  }
+
+  /**
+   * @param directoryPath
+   *          the directoryPath to set
+   */
+  public void setDirectoryPath( String directoryPath ) {
+    this.directoryPath = directoryPath;
+  }
+
+  /**
+   * @return the fileName
+   */
+  public String getFileName() {
+    return fileName;
+  }
+
+  /**
+   * @param fileName
+   *          the fileName to set
+   */
+  public void setFileName( String fileName ) {
+    this.fileName = fileName;
+  }
+
+  /**
+   * @return the transName
+   */
+  public String getTransName() {
+    return transName;
+  }
+
+  /**
+   * @param transName
+   *          the transName to set
+   */
+  public void setTransName( String transName ) {
+    this.transName = transName;
   }
 
   /**
@@ -539,6 +675,36 @@ public class SimpleMappingMeta extends StepWithMappingMeta implements StepMetaIn
    */
   public void setRepository( Repository repository ) {
     this.repository = repository;
+  }
+
+  /**
+   * @return the transObjectId
+   */
+  public ObjectId getTransObjectId() {
+    return transObjectId;
+  }
+
+  /**
+   * @param transObjectId
+   *          the transObjectId to set
+   */
+  public void setTransObjectId( ObjectId transObjectId ) {
+    this.transObjectId = transObjectId;
+  }
+
+  /**
+   * @return the specificationMethod
+   */
+  public ObjectLocationSpecificationMethod getSpecificationMethod() {
+    return specificationMethod;
+  }
+
+  /**
+   * @param specificationMethod
+   *          the specificationMethod to set
+   */
+  public void setSpecificationMethod( ObjectLocationSpecificationMethod specificationMethod ) {
+    this.specificationMethod = specificationMethod;
   }
 
   @Override
@@ -644,5 +810,4 @@ public class SimpleMappingMeta extends StepWithMappingMeta implements StepMetaIn
   public void setOutputMapping( MappingIODefinition outputMapping ) {
     this.outputMapping = outputMapping;
   }
-
 }
